requirements.txt
streamlit==1.40.2
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-texttospeech==2.20.0
googleapis-common-protos==1.65.0
pandas==2.1.1
